+++ examples/ray-tracing/main.rs
@@ -1,16 +1,13 @@
 use std::collections::HashMap;
 use spirq::SpirvBinary;
-use log::info;
 use std::path::Path;
 
 fn main() {
-    env_logger::init();
-
     let spvs = collect_spirv_binaries("assets/effects/ray-tracing");
-    info!("collected spirvs: {:?}", spvs.iter().map(|x| x.0.as_ref()).collect::<Vec<&str>>());
+    println!("collected spirvs: {:?}", spvs.iter().map(|x| x.0.as_ref()).collect::<Vec<&str>>());
     for (name, spv) in spvs {
         let module = spv.reflect_vec().unwrap();
-        info!("{}: {:#?}", name, &module[0]);
+        println!("{}: {:#?}", name, &module[0]);
     }
 }
 
@@ -19,14 +16,12 @@ fn collect_spirv_binaries<P: AsRef<Path>>(path: P) -> HashMap<String, SpirvBinar
     use std::ffi::OsStr;
     use std::fs::{read_dir, File};
     use std::io::Read;
-    use log::warn;
 
     read_dir(path).unwrap()
         .filter_map(|x| match x {
             Ok(rv) => Some(rv.path()),
             Err(err) => {
-                warn!("cannot access to filesystem item: {}", err);
-                None
+                panic!("cannot access to filesystem item: {}", err);
             },
         })
         .filter_map(|x| {